Separate city/data selection into two components
