+++ weather-compare-app/src/App.js
@@ -1,9 +1,18 @@
 import React, { Component } from "react";
+import "./App.css";
+import CitySelection from "./components/CitySelection/CitySelection";
+import DataSelection from "./components/DataSelection/DataSelection";
 import WeatherChart from "./components/WeatherChart/WeatherChart";
 
 class App extends Component {
   render() {
-    return <WeatherChart />;
+    return (
+      <div id="app">
+        <CitySelection/>
+        <WeatherChart/>
+        <DataSelection/>
+      </div>
+    );
   }
 }
