--- weather-compare-app/src/components/CityComparison.css
@@ -1,3 +1,0 @@
-.city-comparison {
-  min-width: 50vw;
-}
